modification de couche services

+++ src/main/java/com/hadjsalem/agencevoyage/services/Impl/DestinationServiceImpl.java
@@ -2,9 +2,12 @@ package com.hadjsalem.agencevoyage.services.Impl;
 import com.hadjsalem.agencevoyage.Common.PageResponse;
 import com.hadjsalem.agencevoyage.dtos.DestinationDto;
 import com.hadjsalem.agencevoyage.entities.Destination;
+import com.hadjsalem.agencevoyage.exceptions.DuplicateEntryException;
 import com.hadjsalem.agencevoyage.mapper.DestinationMapper;
 import com.hadjsalem.agencevoyage.repositories.DestinationRepository;
 import com.hadjsalem.agencevoyage.services.DestinationService;
+import com.hadjsalem.agencevoyage.validators.ObjectsValidators;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -22,6 +25,7 @@ import java.util.Optional;
 public class DestinationServiceImpl implements DestinationService {
     private DestinationRepository destinationRepository;
     private DestinationMapper mapper;
+    private ObjectsValidators<Destination> destinationValidators;
 
 
     @Override
@@ -42,26 +46,44 @@ public class DestinationServiceImpl implements DestinationService {
 
     @Override
     public DestinationDto saveDestination(DestinationDto destinationDto) {
-      Destination Destination1 = mapper.fromdestinationDto(destinationDto);
-      Destination Destination2= destinationRepository.save(Destination1);
-      return mapper.fromdestination(Destination2);
+      Destination destination = mapper.fromdestinationDto(destinationDto);
+      if(destination == null){
+          throw new IllegalArgumentException("destiantion est null");
+      }
+      // Validation des champs
+        destinationValidators.validate(destination);
+      boolean exists = destinationRepository.existsByVille(destinationDto.getVille());
+      if(exists){
+          throw  new DuplicateEntryException(" un destination est existe avec ce nom");
+      }
+      Destination savedDestination = destinationRepository.save(destination);
+      return mapper.fromdestination(savedDestination);
     }
 
     @Override
     public DestinationDto updateDestination(DestinationDto Destinationdto, Long id) {
-      Optional<Destination> Destination1 = destinationRepository.findById(id);
-      if(Destination1.isPresent()){
-      Destination Destination2 = Destination1.get();
-      Destination Destination3= destinationRepository.saveAndFlush(Destination2);
-      return mapper.fromdestination(Destination3);
-      }else {
-          throw  new NoSuchElementException("Destination NotFound");
-      }
+        Optional<Destination> Destination1 = destinationRepository.findById(id);
+        if (Destination1.isPresent()) {
+            Destination destination2 = Destination1.get();
+            destination2.setId(id);
+            destinationValidators.validate(destination2);
+            if (destinationRepository.existsByPays(destination2.getPays())) {
+                throw new DuplicateEntryException("un destination est existe avec cette Pays");
+            }
+            Destination destination3 = destinationRepository.saveAndFlush(destination2);
+            return mapper.fromdestination(destination3);
+        } else {
+            throw new EntityNotFoundException("Destination NotFound");
+        }
     }
-
     @Override
     public void deleteDestination(Long id) {
-       destinationRepository.deleteById(id);
+        if(destinationRepository.findById(id).isPresent()){
+            destinationRepository.deleteById(id);
+        }
+         else {
+             throw new EntityNotFoundException("Destination with this id Not Found");
+        }
     }
 
     @Override
